1.异常信息优化和新增 2.登录判断优化 3.session优化 4.新增webSocket，用于同步展示笔记更改 5.拦截器优化 6.笔记删除逻辑优化 7.新增查询公开几笔接口 8.增加笔记美化、总结、段落润色、图片处理功能 9.七牛云上传位置更改 10.注销账户逻辑优化

+++ src/main/java/top/openfbi/mdnote/note/model/fe/EsNoteResult.java
@@ -9,5 +9,5 @@ public class EsNoteResult {
 
     private List<SimpleNote> noteList;
 
-    private Long totalRecords;
+    private int totalRecords;
 }
